+++ libTicTacToe/RandomGuesser.cpp
@@ -11,7 +11,7 @@ namespace libTicTacToe
 	/// <returns></returns>
 	unsigned int CRandomGuesser::SelectSquare(const std::string state, const std::vector<unsigned int>& availableSquares)
 	{
-		std::uniform_int_distribution<int> distribution(0, availableSquares.size() - 1);
+		std::uniform_int_distribution<size_t> distribution(0, availableSquares.size() - 1);
 		return availableSquares[distribution(_generator)];
 	}
